+++ soc/arm/atmel_sam0/CMakeLists.txt
@@ -3,4 +3,4 @@
 # Copyright (c) 2017 Google LLC.
 # SPDX-License-Identifier: Apache-2.0
 
-add_subdirectory(${SOC_SERIES})
+add_subdirectory(common)
