--- soc/arm/atmel_sam0/samd20/CMakeLists.txt
@@ -1,6 +1,0 @@
-# Makefile - Atmel SAMD20 MCU series.
-#
-# Copyright (c) 2018 Sean Nyekjaer
-# SPDX-License-Identifier: Apache-2.0
-
-zephyr_sources(soc.c)
