--- soc/arm/atmel_sam0/samr21/CMakeLists.txt
@@ -1,6 +1,0 @@
-# Makefile - Atmel SAMR21 MCU series.
-#
-# Copyright (c) 2017 Google LLC.
-# SPDX-License-Identifier: Apache-2.0
-
-zephyr_sources(soc.c)
